+++ libai/models/bert_model.py
@@ -30,8 +30,8 @@ from libai.config import configurable
 
 from .utils import init_method_normal, scaled_init_method_normal
 
-from .build import MODEL_ARCH_REGISTRY
-from .utils import GraphBase, GRAPH_REGISTRY
+from .build import MODEL_ARCH_REGISTRY, GRAPH_REGISTRY
+from .utils import GraphBase
 
 
 class BertExtendedAttnMask(nn.Module):
